jarvis/requirements.txt
google-genai>=1.70.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
playwright>=1.48.0
selenium>=4.25.0
pyautogui>=0.9.54
PyGetWindow>=0.0.9; sys_platform == 'win32'
pyperclip>=1.9.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
psutil>=5.9.0
Pillow>=10.0.0

